+++ tests/testthat/test_mrClosed_MSG.R
@@ -33,6 +33,9 @@ test_that("mrClosed() Single Census messages",{
   mr1 <- mrClosed(ch1)
   expect_error(confint(mr1,conf.level=0),"must be between 0 and 1")
   expect_error(confint(mr1,conf.level=1),"must be between 0 and 1")
+  expect_error(confint(mr1,type="derek"),"should be one of")
+  expect_error(confint(mr1,type="binomial",bin.type="derek"),"should be one of")
+  expect_error(confint(mr1,type="poisson",poi.type="derek"),"should be one of")
 })
   
 test_that("mrClosed Multiple Census errors and warnings",{
@@ -60,4 +63,5 @@ test_that("mrClosed Multiple Census errors and warnings",{
   mr1 <- mrClosed(M1,n1,m1)
   expect_error(confint(mr1,conf.level=0),"must be between 0 and 1")
   expect_error(confint(mr1,conf.level=1),"must be between 0 and 1")
+  expect_error(confint(mr1,poi.type="derek"),"should be one of")
 })